fix: add msg sub to transient message state as side-effect
We have the same issue for subscribing to message events as we had for unsubscribing from them.
Because if we write a subscribe command for a message and reject subscription right after, we will
still have the transient state not reverted. That can cause a message subscription to be created
where it shouldn't be.

As we did for unsubscribing from message subscriptions, we modified the `CatchEventBehavior
.subscribeToMessageEvent()` to update the transient state as a side effect. To conform with that
`DbProcessMessageSubscriptionState.updateToClosingState()` must no longer be adding the message
subscription to the transient state.

Co-authored-by: Nico Korthout <nico.korthout@camunda.com>