feat: Add ad-hoc subprocess builder
Extend the BPMN builder to add an ad-hoc subprocess.

The builder for the ad-hoc subprocess overwrites some internal methods to add elements to the ad-hoc subprocess. The
structure is different to the builder for the embedded subprocess because an ad-hoc subprocess must not contain a start
event. The builders are designed with the assumptions that the elements inside a subprocess are added to a start event
as siblings. The ad-hoc subprocess breaks with this assumption. The current approach is not clean but pragmatic.

(cherry picked from commit 3ab9940a544788fc57f28a9bccb80e049c14c1c6)