fix: update msg sub transient state as side-effect on deletion
To make overall transient state updates consistent, we need to update transient state as side
effect within the engine lifecycle. According to that, ProcessMessageSubscriptionDeleteProcessor
now updates the transient state as a side effect.

Please note that we will retain remaining transient state operations in
DbProcessMessageSubscriptionState for methods:
- onRecovered(): called to fill the transient state on reply
- updateToOpeningState(): called in version migration script
- onSent(): called only in checker

because they are not used within the stream processor lifecycle, so will not cause any race
conditions or inconsistent transient state data.