test: test transient state do not have msg sub on rejection for creation
During process instance migration, it is possible to subscribe to multiple message subscriptions.
While subscribing to first message succeeds, the second could fail due to message name duplication.
The migration command is rejected and all state changes is rolled back including transient state
changes. A test case is added to verify that we rollback the pending subscription in the transient
state and do not subscribe the first message.