feat: Activate ad-hoc activities
When the process instance enters an ad-hoc subprocess, it evaluates the 'activeElementsCollection' expression and
activates all BPMN elements that are part of the resulting list.

If the expression doesn't evaluate to a list or contains strings that are not BPMN element ids, the process instance
creates an incident.

(cherry picked from commit 873ec003eaaa8f7d9da07a2fc70377ec4218040f)