feat: allow to provide scope and record value
We have cases were the element instances doesn't exist yet.
This is especially the case where we want to create the treePath, on
ACTIVATING intent.

Here we can use the flow scope key and ProcessInstanceRecordvalue, to
find all necessary information.

If the flowScopeKey is -1 this means we are at the root.