fix: use the resource loading cache correctly
Break the loading chain, i.e. if a resource is
not found in the cache, look it up directly in
the resource_by_id column family. The current
implementation was creating a loading loop.

Remove eager loading of the cache with lazy
loading, i.e. we only cache a resource if it is
explicitly looked up.

Replace the getIfPresent cache method with a get
method since the former doesn't load the cache if
the resource is not present there.