fix: Activate ad-hoc elements after activation
Correct the order when child elements of the ad-hoc subprocess are activated. Instead of activating the child elements
between ACTIVATING and ACTIVATED, activate the child elements after ACTIVATED.

This order aligns with other subprocesses. The child elements are always activated after the container is activated.

The order change avoids that child instances are activated when an event subscription of a boundary events fails and an
incident is created.