feat: provide the claims map in the broker request object
Instead of a String format token the user info will be provided by a map object inside the broker record.
The string token object is removed from the Authentication record, replaced by the map claims.
Both the gRPC and the REST gateways set into the broker request object the claims map instead of the token