fix: do not close channels on response timeout
Now that we have introduced heartbeats, we no longer need to react on a single request timeout by closing the channel.
This was disruptive because a single missed response would close the channel even when it was actively used for other
requests.

(cherry picked from commit 75c861370798859148ecbf992ee4461be3bae71e)