refactor: AuthorizationCheckBehavior.isAuthorized method
The method now returns Either<RejectionType, Void> instead of a boolean.
If the user is not authorized, that means not enough permission or tenantId not assigned the method returns a rejection.
The rejection type can be UNATHORIZED if the user doesn't have enough permissions, or NOT_FOUND is the tenantId of the
request is not assigned to the provided user.

Some other small refactor like fetching the user from the state only once.
Adjusted all the processors and the test cases to the new return type