fix: return PersistedTenant instead of TenantRecord
There is no need to create a TenantRecord and setting an entity in here. We should return the persisted tenant. This
contains all the tenant data and can be worked with wherever we need to.
The entity key on the record is only for assigning entities to a tenant. Setting this in the getter only causes
confusion.