test: ensure corrections visible to next listeners
Even though the user task corrections only take their effect once all
task listeners have completed and the operation finishes, any subsequent
listener should be able to see the most recent data of the user task.
That includes any corrections made by previous task listeners. The
reason for this is that a task listener should be able to have all data
available that would come into effect if the listener decides to allow
it.

The data is currently made available through custom headers. This could
change in the future when we provide dedicated methods for it. However,
let's just assume that this is the way for now.

Also note that the priority is not available yet in the custom headers
and thus cannot be tested yet here.