test: ensure corrections are persisted finally
When the task listeners all have completed and the operation is
finished, the latest corrections should have taken effect in the user
task itself.

We can achieve this by writing the data in the terminal event. For
complete listeners this is the `COMPLETED` event.

We should take care that for properties that have been corrected
multiple times, we should only consider the latest correction to be the
value to persist. In more complicated words, the latest correction
corrected the previous correction.