refactor: use enhanced switch statement
We have access to enhanced switch statements in the engine code, which
provide multiple improvements over the legacy switch statements. For
example, we don't have to explicitly return at the end of a block, as
the cases don't fall through.

This refactoring was applied by IntelliJ-IDEA. It may be hard to review
by a person. Please note that I did not make any other changes in this
commit other than the automated replacement between switch case styles.