feat: add copy method to user task record
We already have a wrap method for setting values, but that's not a full
copy. We'll have a need for copy soon. To simplify using it, we can
already a convenience method for it.

As there has been some confusion surrounding wrap and copy, I've added
JavaDoc to highlight the difference.

Typically:
- use wrap when you want to set some properties but don't care that
  the same objects/buffers are referenced by the two records. Use this
  when you can as it's faster than copy
- use copy when you need a full copy because you'll be changing both
  records independently. This method is slower than wrap, so only use it
  if you have to.