feat: add method to correct user task attributes
This methods simply overwrites the values that are marked as corrected
attributes, and adds them to the changedAttributes property.

This allows us to slowly expand the changedAttributes property, for
example when multiple task listeners independently correct different
properties and the end result is that multiple properties have changed.
We should then specify the specific changes in the COMPLETED event when
compared to the latest state "before" the action.