fix: use intermediate record after task listener
When a task listener job is completed, we should use the intermediate
state's user task record to continue rather than the current user task
record. The intermediate one could contain corrections from previous
task listeners (not yet, but that will come in a future commit). If we
were to continue with the current user task, then we could lose
previously made corrections.

Note that we have to copy the intermediateUserTaskRecord after
retrieving it from the state. This is because we will later update the
state using a common record referenced from within the state class. This
referenced record was returned to us by the call to
getIntermediateState. When we update (in the CORRECTED applier), the
current value is read again, overwriting our changes. By copying the
record we separate ours from the one used in the state class.