feat: pass job corrections to user task command
When the task listener job is completed with corrections, we need to set
those corrections as changedAttributes on the COMPLETE_TASK_LISTENER
user task command and overwrite those specific attributes.

This allows us to access the corrections in the user task processor
where we will write a CORRECTED event for them.