feat: write corrections in a CORRECTED event
Once a task listener job is completed with a result with corrections,
those corrections are added to the COMPLETE_TASK_LISTENER user task
command as changedAttributes, and the properties themselves will have
been adjusted on that command.

We can now take these and write them to the intermediate user task
record as a CORRECTED event. This will update the intermediate user task
record in the state with those corrections to make it available to
subsequent task listeners.

By changing the intermediate user task record, which is passed along, we
immediately also have these corrections available when finalizing the
action.