feat: collect all changed attributes on finalize
To finalize the action, we need to compare the current user task record
(the state it had before the action started), and the intermediate
userTaskRecord (the last known state after the last task listener
completed). As some of the task listeners may have made corrections to
the user task data, we need to now mark these attributes as changed.

The result of this is that a completing user task will show that both
priority and assignee changed if these were corrected in separate task
listeners, i.e. each in its own CORRECTED event. Note that each of these
CORRECTED events only marked a single changed attribute. But for the
terminal event (COMPLETED), we want to collect all changes.