feat: support assignee in wrapChangedAttributes
wrapChangedAttributes uses updateAttribute internally, but that method
doesn't support setting assignee. This is because the assignee
property cannot be changed by an update request. However, as we can now
correct any of these user task properties, it's best if we can re-use
these methods. We can make that change regardless of what the endpoint
for updating user tasks looks like.