fix: align candidateXList names
The user task record considers the attribute names candidateGroupsList
and candidateUsersList rather than candidateGroups and candidateUsers.

This leads to some difficulties when mapping the JobResultCorrections to
the User Task Record. It also makes it harder for users to understand,
because you could argue that candidateUsers and candidateGroups are
different properties (they existed in the past on user task).

By aligning the names, we can simplify some code in future commits