fix: parse unexpected REST responses in client
If an unexpected response is returned for REST calls, the Java client needs to handle this
in a meaningful way. The client tries to deserialize "application/json" responses into the defined
target type. If this fails, it returns a ProblemDetail instead, containing the original
payload JSON as detail information.

The same routine is used for "application/json" responses that cannot be deserialized into the expected problem details.