feat: added `corrections` to `gateway.proto`
Reasoning for introducing `StringList` message:
There is a need to check if the attributes were changed (in `RequestMapper`).
Protobuf generates helper methods for each field that allow checking if the value was set.
However, that doesn't apply for the repeated values.
In order to be able to determine if `candidateUsersList` and `candidateGroupsList` were set
 I've introduced a wrapper message type.

I am open to the suggestions and assistance here.