test: test migrate active parallel gateway
Parallel gateway can be active if an incident raised on it. Incident might be raised due to
execution listener job type or retries expression evaluation failure. As a part of parallel gateway
migration, we should support migrating parallel gateways with incidents. A test case is added to
verify the behaviour.