fix: only create one CREATED event per form on distribution
Upon deployment CREATE distribution, we only create one CREATED event per form with the defined
resource content per partition. We use the checksum to identify the unqiue form resources instead
of relying on resource names as those can occur multiple times per deployment.

Otherwise, we are creating too many CREATED events for each form resource, one for each other form
with the same name and the content of that other form. This can lead to forms having the wrong
content in the end, depending on which event was processed last for a form.