test: ensure taken flows are retained
This adds a case similar to shouldActivateInclusiveGateway() with some
differences. The main goal is to check that any sequence flow that was
already taken toward an inclusive gateway is retained after the process
instance modification activating that same gateway.

This is important to test because of the implementation of activating an
inclusive gateway through modification: first increment all incoming
sequence flows to the gateway as taken, and then process the activate
element command for the gateway which should decrement all those flows
again. As we increment and decrement we want to avoid forgetting about
any previously taken ones.

We can test this by activating a joining inclusive gateway, for which
already one sequence flow has been taken. On modification, we expect
that it activates even though there is still a token active that can
reach the gateway. When that token eventually reaches the gateway, it
should trigger again, as all incoming sequence flows have been taken
officially now.

(cherry picked from commit c1bcfa46a452e041148d76a40e8b23600b7c141a)