fix: support activating OR gateway via modify
Adds a simple implementation to support activation of INCLUSIVE_GATEWAY
(OR gateway) through activate instruction of Process Instance
Modification.

By adding this line, we increment all incoming sequence flows to the
gateway as taken once. Note that any previously taken flows, will be
simply considered as X+1, where X is the number of times that flow was
already taken previously (increment).

This means that, when the ProcessInstance:ACTIVATE_ELEMENT command for
the gateway is processed, each of the incoming flows have been taken
atleast once and so the command is not rejected.

When we apply the corresponding ELEMENT_ACTIVATING event, the number of
times each incoming sequence flow is decremented again, ensuring that
any previously taken flows are still recorded after modification has
finished.