feat: retrieve username information from the authentication infos
Not all the command might contains the user key.
In the gRPC gateway we are not able to retrieve the user key information, that means that we need to find another way
to get user information from the state without the user key to check user's permissions.
We can use the user's username to fetch the persisted record.

In the gRPC gateway we already decode the user claims and put them into the token that the gateway send to the broker,
within this user claim the is also the username.