fix: reject duplicate setup commands
Both identity setup and deployment reconstruction commands are written
by startup tasks. They contain checks to not write this command when it
is not necessary but these checks are not sufficient, for example in
case the first setup command was written but not processed.

To avoid generating new keys and a failed consistency check when trying
to mark deployments as reconstructed for the second time, we now
double-check the precondition when processing the command and write a
rejection `ALREADY_EXISTS` when the command was a duplicate.