feat: leave groups and roles before deleting user
This writes and applies `EntityRemoved` events such that a user leaves
all groups and roles before it is deleted.

This is helpful for external consumers such that they do not have to
resolve and clean up those relations themselves.

It also fixes a bug where user deletion would previously leave remnants
in the group and role state. Using the proper events makes this more
future-proof too because we don't need to reimplement the cleanup logic.