feat: switch to new task listener event types
Switches to using the new task listener event types internally.

To backward compatibly support the deprecated task listener event types,
we simply map in the UserTaskTransformer to the corresponding new types.

That means that from this commit onwards, all TaskListeners of an
ExecutableUserTask reference the new event types only even if the
process model references the old style.