feat: add all configured users to the IdentitySetupRecord
Instead of only adding a default user we now iterate over all the users in the configuration and add them all to the
IdentitySetupRecord.

Note that I removed the duplication checks. This became too complex to do here, and these checks are duplicated anyway,
as the processor of the command we write here also performs them.