feat: append migrated event for mapped gateways
To retrieve gateways with at least one sequence flow is taken, we iterate through active sequence
flow ids and extract each joining parallel gateway that the sequence flow is connected. Then, each
mapped target gateway retrieved from the mapping instructions. Please note that, validations for
gateway mappings are opted out and will be implemented in a separate PR.

Later, we iterate over the target gateways and append `ELEMENT_MIGRATED` for each of them. While
appending the migrated record, the existing instance record is used to fill in common fields and
updated fields are set afterward.

Please note that a new key is generated for the gateway migration record. It is because we do not
have an instance for the gateway in the state yet. The instance is created only while applying
`ELEMENT_ACTIVATING` event which is not appended until all incoming sequence flows of the parallel
gateway is taken.

Additionally, we use values only from the target gateway. That way, in the applier, we can retrieve
all incoming sequence flows of the target gateway and update the sequence flow ids that are
currently active to contain the new gateway id.

Tests are currently not passing since the applier is not implemented, and it will be implemented in
the next commit.