feat: add decrementNumberOfTakenSequenceFlows to use it later in applier
Currently, we have a method `decrementNumberOfTakenSequenceFlows` to decrement all sequence flows
on the joining gateway (regardless of sequence flow id) during activation of the gateway. However,
with migration, we need to decrement and increment only specific
`flowScope-gatewayId-sequenceFlowId` pairs. The implementation is the same as the existing one but
this time we retrieve the record by its full key instead of prefix.