feat: add visitTakenSequenceFlows to use it later in applier
A method is added to visit each active sequence flow in the flow scope. It is needed to identify
sequence flows exist among the target gateway incoming sequence flows. Then, we can decrement the
count for the old gateway id and increment it for the new gateway id in the applier.