feat: update taken sequence flow gateway ids in the state
In the applier, the number of sequence flows taken for the `flowScopeKey-gatewayId-sequenceFlowId`
pair should be updated to contain the new gateway id. To achieve that, the applier is modified.

First, we make sure there are no instances for the gateway. Because, if there is an element
instance for the gateway, that means the gateway is active but there is an incident raised on it.
That case still needs to be supported. Therefore, we make sure there is no element with the key
(remember we used the key generator to generate a new key) and the type is parallel gateway to
update taken sequence flows only. We then return because there is no instance to migrate for the
gateway so the remaining execution is not needed.

To update taken sequence flows, we first retrieve the incoming sequence flow ids of the target
gateway. Then, for each taken sequence flow in the flow scope, if current sequence flow is
connected to the target gateway, we decrement the count for the old gateway id (delete if 0) and
increment (insert if 0) the count for the new gateway id. Please note that we do that for the
number of times the sequence flow is taken. It is because the old gateway id will no longer exist
on the instance after the migration, so we need to make sure to clean it up from the state.

I directly used the available decrement and increment methods which are updating the count one by
one since the operation is not expected to run for a too high `number` variable.