fix: correctly handle group key during update
Remove the explicit setting of a partition in a
BrokerGroupUpdateRequest. A group is created on
partition 1, and the partition id is encoded in
the group key. It is not necessary to explicitly
set the partition id to the deployment partition
since that ID can already be decoded from the key.

If the ID is explicitly set to 1, and the key
doesn't exist, we prevent the Gateway from detecting
false keys and returning PartitionNotFoundExceptions.