fix: eliminate duplicated snapshot request for indices (not template)
Indices (that were not templates) where included twice in the snapshot
requests, making it impossible to restore them.

Extracted the logic for generating the list of indices to be included in
the requests from BackupServiceImpl to BackupProperties to ease the
testing.

Added a unit test where each list of indices is checked. Comparing this
with the original requirements a discrepancies was found for a couple of
indices.

relates #26066