test: add test to verify migration rejected for updated flow ids
To keep joining gateway migrations simpler, we decided to require taken sequence flow ids to stay
the same in the target process definition. Please see issue comment
https://github.com/camunda/camunda/issues/22067#issuecomment-2520870747 for details.

A test is added to verify the migration is rejected if one of the taken incoming sequence flow ids
of the joining parallel is changed in the target.