feat: validate migration for updated sequence flow ids
To keep joining gateway migrations simpler, we decided to require taken sequence flow ids to stay
the same in the target process definition. Please see issue comment
https://github.com/camunda/camunda/issues/22067#issuecomment-2520870747 for details.

`ProcessInstanceMigrationMigrateProcessor.getMappedGateways()` method is updated to validate each
active (taken) sequence flow to exist in the target. `The distinct()` call is removed because we
need to validate each sequence flow even it they are connected to the same source gateway. At the
end of the method, we still convert result to Set in order to eliminate duplicated `targetGateway`
objects.

In ProcessInstanceMigrationPreconditions, a method is created to verify if the taken sequence flow
is connected to the target gateway. If not, we reject the migration.