refactor: simpler migration start and completion
This rewrites a bit how migrations are started and awaited. Instead of
using custom `ApplicationEvent`s, we now just declare an
`ApplicationRunner` that runs all available `Migrator`s. If any of them
fail, the containing Spring application fails to start.

Defining an `ApplicationRunner` makes it easier to wait for successful
migration in tests.