refactor: write migrated event for sequence flow
We decided to write `ELEMENT_MIGRATED` event for the sequence flow instead of the joining gateway.
A joining gateway does not have an instance in the state until all incoming sequence flows are
taken. Sequence flows do not have an instance in the state. But, on `SEQUENCE_FLOW_TAKEN` event,
all data related to the sequence flow is created. Additionally, the joining gateway will not have a
key yet while we are migrating the instance. Additionally, in the applier, we are actually updating
the same column family as we do while applying `SEQUENCE_FLOW_TAKEN` event.

In short, while migrating joining gateways, sequence flow data exist in the state but the joining
gateway instance is not. Therefore, we chose to write migrated event for sequence flows.