refactor: add integrity check for active sequence flow ids
There is a warning on the `getActiveSequenceFlowIds()` method about usage:
Warning, this method should not be used for process instances created before 8.6. It may provide
incorrect information for such process instances.

To make sure we have complete active sequence flow ids, we need to make sure active sequence flow
ids count matches with the active sequence flow count that exists before 8.6. Additionally,
`size()` method is added to `ArrayProperty` to improve the performance of the count query.