refactor: use visitTakenSequenceFlows to retrieve sequence flow ids
Previously we were using getSequenceFlowIds over the elementInstance but the result of that method
could be invalid for process instances created before 8.6. Instead, we can use your newly created
method ElementInstanceState.visitTakenSequenceFlows to visit them. This provides all the needed
info reliably as it is build around the numberOfTakenSequenceFlowsColumnFamily which has been the
state tracking this for the joining parallel gateway since "forever". This would also mean we don't
need a safety check.

Note: `getSequenceFlowsToMigrate()` method will be simplified along with validations PR.