test: test migrate active inclusive gateway
Inclusive gateway can be active if an incident raised on it. Incident might be raised due to
execution listener job type or retries expression evaluation failure. As a part of inclusive gateway
migration, we should support migrating inclusive gateways with incidents. A test case is added to
verify the behaviour.