fix: persist entire assigned user task
We need to persist any changed data of the user task. Rather than adding
complexity here using the changed attributes, we can simply persist the
record as it was written in the event.

Note that we have to remove any variables, action, and changed
attributes as these are not to be stored long term.

- variables are available on the complete command only and later in the
  creating and created events, the variables are empty, they have been
  written to an EventTrigger event to become process variables
- action was only relevant to the task listeners and the final event
  (so the action shows what happened), but is not necessary to persist
  afterward
- changed attributes only show what attributes have changed. When
  assigning x, it should at least contain x, but also any corrections.
  Like action, the changed attributes are no longer relevant afterward.