fix: set assignee changed on assigning
When assigning (or claiming), the assignee may change compared to
previously. If it changes, we should track this in the changedAttributes
property of the ASSIGNING/CLAIMING and the ASSIGNED events.

This became clear when trying out a different implementation for the
`ASSIGNED` event applier, where the changed attributes are used to
persist the changes.

Note that the corrections already adjust the changedAttributes when the
correction creates a diff.

I've moved the setter for the assignee down under the equals check
otherwise the equals check wouldn't find any differences.