fix: decouple the persisted user task from state
In order to call `userTaskState.getUserTask` from within the ASSIGNED
applier, the record must be copied. Otherwise, the read overwrites the
user task passed to the applier. This happens because previously the
user task retrieved from the state was passed around as a reference to
a field existing in the state class. Any call to the state class that
modifies this field directly takes effect on the passed around
reference. We can avoid this by copying the record.