feat: reject migration if all incoming flows of the gateway are taken
While processing commands, `ACTIVATE_ELEMENT` command might not fit into the existing command batch
and can be deferred to be processed later. In the meantime, if an instance migration command is
executed, the migrate command will end up changing the instance that is not stable. E.g. the
gateway should be activated/completed but the execution is deferred.

If we do not prevent such migration execution, it will lead process instance records will be
written to the log stream and processed with wrong process definition key, bpmn process id or other
process related information. That breaks the data integrity of the log stream.

To prevent that, we need to make sure the joining parallel gateway is in a stable state. E.g.
incoming sequence flows to it is only partially taken, so valid for migration.