refactor: rename test case
This goal of this test is verify we can send a command over REST and that all the authorization in between works as
expected. It doesn't matter that we use a deployment to test this. All that matters is that we don't end up with a
FORBIDDEN response, unless we don't provide the user with the required permissions.