fix: allow explicitly passed null values
As our API is clear about the different forms of usage, allowing null is
important. When null is used, the property will not be corrected
(ignored).

This is especially essential when the calls to for example
`correctAssignee` are made using dynamically calculated variables or
passed in parameters. If `null` is passed to it, the behavior should be
the same.

With this commit, we now support passing null explicitly to the correctX
methods.

Co-authored-by: ana.vinogradova <ana.vinogradova@camunda.com>