feat: allow setting corrections as object
This can be useful when reusing the corrections object or when it's
passed in as a parameter to a method.

Two methods are introduced:
- CompleteJobCommandStep.correct(Corrections)
- CompleteJobResult.correct(Corrections)

This required us to refactor the CompleteJobResult by moving the
corrections out into a separate class.

Note that the correct(Corrections) method allows setting null, but the
internal implementation requires an instance of JobResultCorrections.
This is why the correct method creates a new instance of it when null is
passed to it.

We did not add tests for CompleteJobCommandStep.correct(Corrections)
because it's used directly by withResult(CompleteJobResult), as one is
just a convenience method for the other.

Co-authored-by: ana.vinogradova <ana.vinogradova@camunda.com>