feat: add convenience lambda method for withResult
Allows easily:
- modifying the previously made result changes
- setting without knowing how to instantiate the class

We deliberately chose to set denied to false, because the
CompleteJobResult class uses boolean rather than Boolean.
The impact of this is that the denied flag will always be sent to the
API when this method is used. That should be fine because we say that
the default is false anyways.

Co-authored-by: ana.vinogradova <ana.vinogradova@camunda.com>