refactor: statically import assertThat
While we're generally against static method imports (see
ZeebeCodeStyle), this is a place where it makes sense to do so.

In the following commit, we'll use assertThat from
io.camunda.zeebe.protocol.record.Assertions. Note that there already was
usage of this as well. However, we also use
org.assertj.core.api.Assertions. The easiest way to use both is to allow
method overloading to do it's thing. We can achieve this by statically
importing both methods.

It's arguable that assertThat is common enough in our test code that
everyone understands that this concerns one of these methods and that
the right one will be used due to overloading.