test: ensure task listeners can be migrated
When a user task has an active task listener (job), we should ensure
that the task and the associated job can be migrated correctly on
instance migration.

The test could be added and run successfully without any production code
changes, because it already works out of the box. This is due to the
generic solution to migrate any jobs associated to element instances.
The task listener job is associated to the user task instance.