refactor: use returned command rejection
The call to `.migrate()` returns the corresponding event or command
rejection. As the method is set up to `.expectRejection()`, it will
return a command rejection. This means there is no need to look up the
command rejection separately again using the RecordingExporter directly.

Instead, we can simply use the returned object.