refactor: rename variable for correctness
flowScopeKey is easy to confuse with flowScopeInstanceKey, because flow
scope is sometimes used interchangeably between the parent element (flow
scope) and the scope of an operation in the flow (flow scope).

To avoid confusion in this method, this key is specifically the key of
the element instance, i.e. the elementInstanceKey. That it is used as
the flowScopeKey in the .visitTakenSequenceFlows method is not relevant
in the context of this method.

For example, on line 480, the elementInstanceKey refers to the flow
scope (parent) of the sourceGateway.

If this seems confusing, we could consider explaining that the source
gateway is actually a child of the current element instance.