fix: in NettyMessaging service heartbeats are sent as req/reply based on role
Heartbeats are now initiated from the client, which sends a
ProtocolRequest message with subject "internal-heartbeat" whenever the
channel has been idle for too much.
The server, upon receive such message, will send a ProtocolReply
response with the same id as the request.
The client is then able to diagnose missed heartbeats and close the
channel if he does not receive a reply in time.

If the server has received at least one heartbeat, but it did not
receive an heartbeat for more than the specified timeout, it will close
the connection.
This will avoid closing the connection when a client does not support
heartbeats yet (i.e. a previous version)

closes #26602