feat: define full user object response in spec
Define a full user object for the response of Create User
endpoint. Since responses where the user key is a number
will be deprecated with the upcoming release, a base
response is defined that contains all user properties
except the user key. This base response is extended by
the existing UserCreateResponse and UserCreateResult
objects.