feat: client does not close the connection is server does not reply to heartbeats
Similarly to what was done in the server, if the client never receives a
reply to its heartbeats it will not close the connection, assuming the
server does not support this functionality yet.

When the server supports this functionality, all ProtocolReply are used
as an indication that the TCP is working.

relates #26602