test: add test to verify migration rejected for updated flow ids
To keep joining gateway migrations simpler, we decided to require taken sequence flow ids to stay
the same in the target process definition. Please see issue comment
#22067 (comment) for details.

A test is added to verify the migration is rejected if one of the taken incoming sequence flow ids
of the joining inclusive gateway is changed in the target.