test: reject migration if target gateway has less incoming seq flows
We will reject migration if target gateway has less incoming sequence flows than the source
gateway. There are two main reasons for that:
- Improved UX: if a gateway with 3 incoming flows and 2 are taken is mapped to a gateway with 2
incoming sequence flows in total. Users will expect the target gateway to complete automatically
after the migration. But this is not the case with migration and that could cause ambiguity in the
instance migration behaviour.
- C7 feature alignment: the validation also exists in C7 instance migration API (see
https://docs.camunda.org/manual/7.22/user-guide/process-engine/process-instance-migration/
#inclusive-parallel-gateway), so we will have alignment between features when users move from C7 to
C8.