feat: track "assignee" in `changedAttributes` after user task creation
This change updates the `userTaskAssigning` method to ensure that the "assignee" value is added
to the `changedAttributes` list in the user task record when the task is created with a defined
`assignee` property.

By including the "assignee" as a changed attribute, the subsequent `ASSIGNING` and `ASSIGNED`
events triggered after the `CREATED` event will reflect the change accurately.