refactor: rely on `changedAttributes` when handling user task assignment and updates
This update refactors the `UserTaskHandler` class to rely solely on the `changedAttributes` list
from the user task record value when processing user task assignment and updates. With Zeebe now
including "assignee" as a changed attribute in the `ASSIGNED` record, manually adding "assignee"
to the entity's `changedAttributes` upon handling the `ASSIGNED` intent is no longer necessary.

As a result, the handling logic for `ASSIGNED` and `UPDATED` intents is unified, allowing both to
use the `updateChangedAttributes` method for processing, reducing redundancy.

This refactoring also resolves an issue where a user task correction during assignment could result
in duplicate "assignee" entries in the `changedAttributes` list. Previously, the `handleAssignment`
method would redundantly add "assignee" to entity `changedAttributes` even when it was already present in
the record value `changedAttributes`, as described in https://github.com/camunda/camunda/issues/26791.